Fikser posisjon for både kapittelnavigasjon og lenkeliste
