+++ src/components/parts/_legacy/menu-list/MenuList.tsx
@@ -31,7 +31,7 @@ export const MenuList = (
 
     return (
         validItemKeys.length > 0 && (
-            <div className={style.menuList}>
+            <div className={style.menuList} id="link-list-container">
                 {validItemKeys.map((key) => {
                     const { links } = menuListItems[key];
                     const isOpen =
